+++ src/api/java/mekanism/api/fluid/ISidedFluidHandler.java
@@ -2,9 +2,9 @@ package mekanism.api.fluid;
 
 import mekanism.api.Action;
 import mekanism.api.annotations.NothingNullByDefault;
+import mekanism.api.fabric.transfer.fluids.IFluidHandler;
 import net.minecraft.core.Direction;
-import net.neoforged.neoforge.fluids.FluidStack;
-import net.neoforged.neoforge.fluids.capability.IFluidHandler;
+import io.github.fabricators_of_create.porting_lib.fluids.FluidStack;
 import org.jetbrains.annotations.Nullable;
 
 /**
@@ -93,10 +93,10 @@ public interface ISidedFluidHandler extends IExtendedFluidHandler {
      *
      * @return The maximum fluid amount held by the tank.
      */
-    int getTankCapacity(int tank, @Nullable Direction side);
+    long getTankCapacity(int tank, @Nullable Direction side);
 
     @Override
-    default int getTankCapacity(int tank) {
+    default long getTankCapacity(int tank) {
         return getTankCapacity(tank, getFluidSideFor());
     }
 
@@ -150,7 +150,7 @@ public interface ISidedFluidHandler extends IExtendedFluidHandler {
     }
 
     /**
-     * A sided variant of {@link IExtendedFluidHandler#extractFluid(int, int, Action)}, docs copied for convenience.
+     * A sided variant of {@link IExtendedFluidHandler#extractFluid(int, long, Action)}, docs copied for convenience.
      * <p>
      * Extracts a {@link FluidStack} from a specific tank in this handler.
      * <p>
@@ -165,10 +165,10 @@ public interface ISidedFluidHandler extends IExtendedFluidHandler {
      * @return {@link FluidStack} extracted from the tank, must be empty if nothing can be extracted. The returned {@link FluidStack} can be safely modified after, so the
      * tank should return a new or copied stack.
      */
-    FluidStack extractFluid(int tank, int amount, @Nullable Direction side, Action action);
+    FluidStack extractFluid(int tank, long amount, @Nullable Direction side, Action action);
 
     @Override
-    default FluidStack extractFluid(int tank, int amount, Action action) {
+    default FluidStack extractFluid(int tank, long amount, Action action) {
         return extractFluid(tank, amount, getFluidSideFor(), action);
     }
 
@@ -197,7 +197,7 @@ public interface ISidedFluidHandler extends IExtendedFluidHandler {
     }
 
     /**
-     * A sided variant of {@link IExtendedFluidHandler#extractFluid(int, Action)}, docs copied for convenience.
+     * A sided variant of {@link IExtendedFluidHandler#extractFluid(long, Action)}, docs copied for convenience.
      * <p>
      * Extracts a {@link FluidStack} from this handler, distribution is left <strong>entirely</strong> to this {@link IExtendedFluidHandler}.
      * <p>
@@ -215,7 +215,7 @@ public interface ISidedFluidHandler extends IExtendedFluidHandler {
      * extracted is found, all future extractions will make sure to also make sure they are for the same type of fluid.
      * @apiNote It is not guaranteed that the default implementation will be how this {@link IExtendedFluidHandler} ends up distributing the extraction.
      */
-    default FluidStack extractFluid(int amount, @Nullable Direction side, Action action) {
+    default FluidStack extractFluid(long amount, @Nullable Direction side, Action action) {
         return ExtendedFluidHandlerUtils.extract(amount, side, action, this::getTanks, this::getFluidInTank, this::extractFluid);
     }
